// frontend/src/apisv2beta1/run/models/V2beta1TaskScope.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Describes the runtime state of an entity.
 *
 *  - RUNTIME_STATE_UNSPECIFIED: Default value. This value is not used.
 *  - PENDING: Service is preparing to execute an entity.
 *  - RUNNING: Entity execution is in progress.
 *  - SUCCEEDED: Entity completed successfully.
 *  - SKIPPED: Entity has been skipped. For example, due to caching.
 *  - FAILED: Entity execution has failed.
 *  - CANCELING: Entity is being canceled. From this state, an entity may only
 * change its state to SUCCEEDED, FAILED or CANCELED.
 *  - CANCELED: Entity has been canceled.
 *  - PAUSED: Entity has been paused. It can be resumed.
 *  - The propagation scope for Clear and Mark Success actions.
 * @export
 */

export const V2beta1TaskScope = {
  TASK_ONLY: 'TASK_ONLY',
  DOWNSTREAM: 'DOWNSTREAM',
  UPSTREAM: 'UPSTREAM',
  UPSTREAM_DOWNSTREAM: 'UPSTREAM_DOWNSTREAM',
} as const;
export type V2beta1TaskScope = (typeof V2beta1TaskScope)[keyof typeof V2beta1TaskScope];

export function V2beta1TaskScopeFromJSON(json: any): V2beta1TaskScope {
  return json as V2beta1TaskScope;
}

export function V2beta1TaskScopeToJSON(value?: V2beta1TaskScope | null): any {
  return value as any;
}